move docker project to main solution
