--- AowDocker/Program.cs
@@ -1,89 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using Discord;
-using Discord.Commands;
-using Discord.WebSocket;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.EntityFrameworkCore;
-
-namespace AowContainer
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            new Program().RunBotAsync().GetAwaiter().GetResult();
-            
-            //*Sheet test:
-            //SheetQuerry sheetQuerry = new SheetQuerry();
-            //sheetQuerry.SelectSheet();
-
-        }
-        
-        private DiscordSocketClient _client;
-        private CommandService _commands;
-        private IServiceProvider _services;
-
-        public async Task RunBotAsync()
-        {
-            var discordSocketConfig = new DiscordSocketConfig();
-            discordSocketConfig.GatewayIntents = GatewayIntents.All;
-
-            _client = new DiscordSocketClient(discordSocketConfig);
-            _commands = new CommandService();
-
-            _services = new ServiceCollection()
-                .AddSingleton(_client)
-                .AddSingleton(_commands)
-                .BuildServiceProvider();
-            string token = Environment.GetEnvironmentVariable("DISCORD_BOT_TOKEN");
-
-            _client.Log += _client_Log;
-
-            await RegisterCommandsAsync();
-
-            await _client.LoginAsync(TokenType.Bot, token);
-
-            await _client.StartAsync();
-
-            await Task.Delay(-1);
-
-
-
-
-        }
-
-        private Task _client_Log(LogMessage arg)
-        {
-            
-            Console.WriteLine(arg);
-            return Task.CompletedTask;
-        }
-
-        public async Task RegisterCommandsAsync()
-        {
-            _client.MessageReceived += HandleCommandAsync;
-            await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
-        }
-
-        private async Task HandleCommandAsync(SocketMessage arg)
-        {
-            var message = arg as SocketUserMessage;
-            var context = new SocketCommandContext(_client, message);
-            if (message.Author.IsBot) return;
-
-            int argPos = 0;
-            if (message.HasStringPrefix("!", ref argPos))
-            {
-                var result = await _commands.ExecuteAsync(context, argPos, _services);
-                if (!result.IsSuccess) Console.WriteLine(result.ErrorReason);
-                if (result.Error.Equals(CommandError.UnmetPrecondition)) await message.Channel.SendMessageAsync(result.ErrorReason);
-            }
-        }
-    }
-}
