--- AowDocker/Table/Guild.cs
@@ -1,17 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Text;
-
-namespace Table
-{
-    public class Guild
-    {
-        [Key]
-        public long guild { get; set; }
-        public long discordid { get; set; }
-        public string startdate { get; set; }
-        public int exemption { get; set; }
-    }
-}
