--- AowDocker/Table/Person_info.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Table
-{
-    public class Person_info
-    {
-        public long id { get; set; }
-        public string name { get; set; }
-    }
-}
