+++ DoD.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.3.32922.545
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DoD", "DoD\DoD.csproj", "{C71E7D81-7F55-4D10-BBD0-2FF4B0FAA48B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AowDocker", "AowDocker\AowDocker.csproj", "{C4CDF3D3-8A66-4580-9A0D-35B2F9D7921F}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{C71E7D81-7F55-4D10-BBD0-2FF4B0FAA48B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C71E7D81-7F55-4D10-BBD0-2FF4B0FAA48B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C71E7D81-7F55-4D10-BBD0-2FF4B0FAA48B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C4CDF3D3-8A66-4580-9A0D-35B2F9D7921F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C4CDF3D3-8A66-4580-9A0D-35B2F9D7921F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C4CDF3D3-8A66-4580-9A0D-35B2F9D7921F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C4CDF3D3-8A66-4580-9A0D-35B2F9D7921F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
