+++ DoD/Methods/SheetQuerry.cs
@@ -12,6 +12,9 @@ using Discord.Commands;
 using Discord.WebSocket;
 using DoD;
 using Google.Apis.Sheets.v4.Data;
+using Table;
+using Context;
+using System.Net;
 
 namespace DoD
 {
@@ -24,7 +27,8 @@ namespace DoD
             {"VAL", "1dpBYsLLBTaOY5XPTQIdUwvauTwB8h0rNGRIWkXdc_HU"},
             {"SAO", "1n8YMi0_sb1hXphbl1wgO9uNvyHnIVLbWPSBDEJk6GAQ"},
             {"FTW", "1QXIFV40VS1fjpyivxcneQ3aCWcmpv8ji8wyouj7fIiM"},
-            {"SHR", "1rodipreQnXxR76qDeg9LzKb_SNp5aFllmukQF7-sbfk"}
+            {"SHR", "1rodipreQnXxR76qDeg9LzKb_SNp5aFllmukQF7-sbfk"},
+            {"DOM", "1nGxsE2XEZKjRV4voIHNpIIccEjw1yUZOS1UsQuXnQRY"}
         };
         public static string SpreadsheetId { get; set; }
         static readonly string[] Scopes = { SheetsService.Scope.Spreadsheets };
@@ -37,7 +41,7 @@ namespace DoD
         {
             SpreadsheetId = sheetid[$"{guild}"];
             GoogleCredential credential;
-            using (var stream = new FileStream("credentials.json", FileMode.Open, FileAccess.Read))
+            using (var stream = new FileStream("/source/sheet/credentials.json", FileMode.Open, FileAccess.Read))
             {
                 credential = GoogleCredential.FromStream(stream)
                     .CreateScoped(Scopes);
@@ -114,7 +118,7 @@ namespace DoD
             }
             else if (guild == "VAL")
             {
-                var range = $"{sheet}!A{max}:L5757";
+                var range = $"{sheet}!A{max}:J5757";
                 var request = service.Spreadsheets.Values.Get(SpreadsheetId, range);
                 var response = request.Execute();
                 var values = response.Values;
@@ -138,7 +142,7 @@ namespace DoD
                             lineid = Int32.Parse((string)row[0]),
                             name = (string)row[1],
                             date = (string)row[2],
-                            guild = "",
+                            guild = guild,
                             type = (string)row[3],
                             food = Double.Parse((string)row[4]),
                             parts = Double.Parse((string)row[5]),
@@ -183,7 +187,7 @@ namespace DoD
                             lineid = Int32.Parse((string)row[0]),
                             name = (string)row[1],
                             date = (string)row[2],
-                            guild = "",
+                            guild = guild,
                             type = (string)row[5],
                             food = Double.Parse((string)row[6]),
                             parts = Double.Parse((string)row[7]),
@@ -191,6 +195,51 @@ namespace DoD
                             gas = Double.Parse((string)row[9]),
                             cash = Double.Parse((string)row[10]),
                             shadow = Double.Parse((string)row[11]),
+                        };
+                        context.data_bank.Add(std);
+                        context.SaveChanges();
+                    }
+                    then = DateTime.Now.AddHours(1);
+                }
+                else
+                {
+                    Console.WriteLine("No data found");
+                }
+            }
+            else if (guild == "DOM")
+            {
+                var range = $"{sheet}!A{max}:J5757";
+                var request = service.Spreadsheets.Values.Get(SpreadsheetId, range);
+                var response = request.Execute();
+                var values = response.Values;
+                if (values != null && values.Count > 0)
+                {
+                    foreach (var row in values)
+                    {
+                        Console.WriteLine($"{row[0]} | {row[1]} | {row[2]} | {row[3]} | {row[4]} | {row[5]} | {row[6]} | {row[7]} | {row[8]} | {row[9]}");
+                        List<Person_info> temp = new List<Person_info>();
+                        temp = dbmethod.SelectPersonName(guild, (string)row[1]);
+                        long? id = null;
+                        if (temp.Count > 0)
+                        {
+                            id = temp[0].id;
+                        }
+                        var context = new DbContext();
+                        DbContext.dbname = guild;
+                        var std = new Data_bank()
+                        {
+                            id = id,
+                            lineid = Int32.Parse((string)row[0]),
+                            name = (string)row[1],
+                            date = (string)row[2],
+                            guild = guild,
+                            type = (string)row[3],
+                            food = Double.Parse((string)row[4]),
+                            parts = Double.Parse((string)row[5]),
+                            electric = Double.Parse((string)row[6]),
+                            gas = Double.Parse((string)row[7]),
+                            cash = Double.Parse((string)row[8]),
+                            shadow = Double.Parse((string)row[9]),
                         };
                         context.data_bank.Add(std);
                         context.SaveChanges();
