+++ DoD/Table/Color.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
 
-namespace DoD
+namespace Table
 {
     public class Color
     {
