+++ DoD/Table/Data_bank.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace DoD
+namespace Table
 {
     public class Data_bank
     {
