+++ DoD/Table/Guild.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
 
-namespace DoD
+namespace Table
 {
     public class Guild
     {
