+++ DoD/Table/Person_info.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DoD
+namespace Table
 {
     public class Person_info
     {
